jest settime 수정, story_utils 생성
